refactor: 将显示设置相关组件就近归入 display 目录，提升设置页模块内聚性
