+++ src/renderer/pages/settings/display/components/ThemeSwitcher.tsx
